refactor(generation): modularize text generation with tools and streamline utility functions
- Moved the `generateTextWithTools` and `streamWithTools` functions into a new `generationWithTools.ts` module, enhancing modularity and separation of concerns.
- Introduced the `buildToolSet` function to encapsulate tool set construction, improving code organization and readability.
- Extracted the `buildGenerationSettings` function into a dedicated `generationHelpers.ts` file, promoting reusability and clarity in model options construction.
