+++ packages/core/src/index.ts
@@ -27,3 +27,4 @@ export * from "./ragknowledge.ts";
 export * from "./utils.ts";
 export * from "./tokenTrimming.ts";
 export * from "./imgGeneration.ts";
+export * from "./generationWithTools.ts";
